# Conflicts:
#	src/AGIL/DefaultBundle/Resources/config/services.yml
#	web/js/tags.js

+++ src/AGIL/DefaultBundle/Controller/TagController.php
@@ -3,6 +3,7 @@
 namespace AGIL\DefaultBundle\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -21,6 +22,9 @@ class TagController extends Controller {
 	 * Récupère une liste de tags dont le préfixe est $char et la renvoie au format JSON
 	 */
 	public function searchAction(Request $request) {
+		if (!$request->isXmlHttpRequest()) {
+			throw $this->createNotFoundException('Ressource indisponible');
+		}
 		// On récupère la valeur envoyée par la requête
 		$prefix = $request->request->get('prefix');
 
@@ -35,9 +39,8 @@ class TagController extends Controller {
 		foreach ($tagsList as $tag) {
 			$jsonTagsList[] = $tag->getTagName();
 		}
-
 		// Retourne le tableau encodé en JSON
-		return new Response(json_encode($jsonTagsList));
+		return new JsonResponse($jsonTagsList);
 	}
 
 	/**
@@ -46,19 +49,22 @@ class TagController extends Controller {
 	 * Récupère un tag dans la base de données pour le supprimer
 	 */
 	public function removeAction(Request $request) {
+		if (!$request->isXmlHttpRequest()) {
+			throw $this->createNotFoundException('Ressource indisponible');
+		}
 		$tagName = $request->request->get('tagName');
 
 		$em = $this->getDoctrine()->getManager();
 
 		$tag = $em->getRepository('AGILDefaultBundle:AgilTag')
 			// On peut utiliser le OneBy car chaque tag est unique
-				// méthode magique
-				->findOneByTagName($tagName);
+			// méthode magique
+			->findOneByTagName($tagName);
 
 		if (null !== $tag) {
 			$em->remove($tag);
 			$em->flush();
-			return new Response();
 		}
+		return new Response();
 	}
 }
